Fix More

+++ shared/src/map/camera/MapCamera3DHelper.h
@@ -29,20 +29,20 @@ public:
         return cameraDistance * tan((fovy * 0.5) * (M_PI / 180.0));
     }
 
-    static Vec3D raySphereIntersection(const Vec3D& rayStart, const Vec3D &rayEnd, const Vec3D &sphereCenter, float radius, bool &didHit) {
+    static Vec3D raySphereIntersection(const Vec3D& rayStart, const Vec3D &rayEnd, const Vec3D &sphereCenter, double radius, bool &didHit) {
 
         // https://www.khoury.northeastern.edu/home/fell/CS4300/Lectures/Ray-TracingFormulas.pdf
-        float dx = rayEnd.x - rayStart.x;
-        float dy = rayEnd.y - rayStart.y;
-        float dz = rayEnd.z - rayStart.z;
+        double dx = rayEnd.x - rayStart.x;
+        double dy = rayEnd.y - rayStart.y;
+        double dz = rayEnd.z - rayStart.z;
 
-        float cx = sphereCenter.x;
-        float cy = sphereCenter.y;
-        float cz = sphereCenter.z;
+        double cx = sphereCenter.x;
+        double cy = sphereCenter.y;
+        double cz = sphereCenter.z;
 
-        float x0 = rayStart.x;
-        float y0 = rayStart.y;
-        float z0 = rayStart.z;
+        double x0 = rayStart.x;
+        double y0 = rayStart.y;
+        double z0 = rayStart.z;
 
         auto a = dx*dx + dy*dy + dz*dz;
         auto b = 2.0 * dx * (x0 - cx) + 2.0 * dy * (y0 - cy) + 2.0 * dz * (z0 - cz);
